requirements.txt
Flask==2.0.3
gunicorn==20.1.0
Jinja2==3.0.3
Werkzeug==2.0.3
Pillow==10.3.0
requests==2.31.0